chore: run connect and echo on all key types, muxers and encrypters
