+++ src/streams/index.ts
@@ -1,7 +1,7 @@
+import { runTests } from '../utils/test-matrix.js'
 import { echoStreamTests } from './echo.js'
 import type { DaemonFactory } from '../index.js'
 
 export async function streamTests (factory: DaemonFactory): Promise<void> {
-  echoStreamTests(factory, 'mplex')
-  echoStreamTests(factory, 'yamux')
+  runTests('echo', echoStreamTests, factory)
 }
